[!!!][TASK] Use Base64CryptoManager for transmission "security"
The hardcoded RC4 cipher that previously was used in OpenSSLCryptoManager is not available in OpenSSL 3.0 and onwards. As a temporary measure, we are switching to the deprecated Base64CryptoManager. This is fine, because Caretaker stems from a time in which transport security wasn't a widespread thing, and securing data in transport was mandatory. Nowadays, with HTTPS being a standard everywhere, this level of transport security with additional encryption is not required anymore.

This is a MASSIVE breaking change because just using this version with a current Caretaker server breaks communications entirely. The matching client version in the server needs a release to use the Base64 crypto manager as well, so that communication can be properly de/encoded between instances.

Related: https://projekte.in2code.de/issues/64998